Move some classes out.
