+++ inputs/__init__.py
@@ -43,25 +43,21 @@ from __future__ import print_function
 from __future__ import division
 
 import os
-import sys
 import io
 import glob
 import struct
-import platform
-import math
 import time
-import codecs
 from warnings import warn
-from itertools import count
-from operator import itemgetter
-from multiprocessing import Process, Pipe
 import ctypes
+
+from .platforms.baselistener import BaseListener
+from .devices.gamepad._win import XinputState
+from .devices.device import InputDevice
 
 from .constants import (
     XINPUT_DLL_NAMES,
     XINPUT_ERROR_DEVICE_NOT_CONNECTED,
     XINPUT_ERROR_SUCCESS,
-    EVENT_TYPES,
     WIN_KEYBOARD_CODES,
     WIN_MOUSE_CODES,
     MAC_EVENT_CODES,
@@ -74,73 +70,20 @@ from .constants import (
 
 from .errors import (
     PERMISSIONS_ERROR_TEXT,
-    UnpluggedError,
-    NoDevicePath,
     UnknownEventType,
     UnknownEventCode
 )
 
+from .platforms.system import WIN, MAC, NIX
+from .platforms.c import DWORD, HANDLE, WPARAM, LPARAM, MSG, EVENT_FORMAT, EVENT_SIZE, convert_timeval
+from .devices.gamepad.gamepad import GamePad
 
 __version__ = "0.6"
 
 
-WIN = True if platform.system() == "Windows" else False
-MAC = True if platform.system() == "Darwin" else False
-NIX = True if platform.system() == "Linux" else False
-
-if WIN:
-    # pylint: disable=wrong-import-position
-    import ctypes.wintypes
-
-    DWORD = ctypes.wintypes.DWORD
-    HANDLE = ctypes.wintypes.HANDLE
-    WPARAM = ctypes.wintypes.WPARAM
-    LPARAM = ctypes.wintypes.WPARAM
-    MSG = ctypes.wintypes.MSG
-else:
-    DWORD = ctypes.c_ulong
-    HANDLE = ctypes.c_void_p
-    WPARAM = ctypes.c_ulonglong
-    LPARAM = ctypes.c_ulonglong
-    MSG = ctypes.Structure
-
 if NIX:
     from fcntl import ioctl
 
-OLD = sys.version_info < (3, 4)
-
-# Standard event format for most devices.
-# long, long, unsigned short, unsigned short, int
-EVENT_FORMAT = str("llHHi")
-
-EVENT_SIZE = struct.calcsize(EVENT_FORMAT)
-
-
-def chunks(raw):
-    """Yield successive EVENT_SIZE sized chunks from raw."""
-    for i in range(0, len(raw), EVENT_SIZE):
-        yield struct.unpack(EVENT_FORMAT, raw[i : i + EVENT_SIZE])
-
-
-if OLD:
-
-    def iter_unpack(raw):
-        """Yield successive EVENT_SIZE chunks from message."""
-        return chunks(raw)
-
-else:
-
-    def iter_unpack(raw):
-        """Yield successive EVENT_SIZE chunks from message."""
-        return struct.iter_unpack(EVENT_FORMAT, raw)
-
-
-def convert_timeval(seconds_since_epoch):
-    """Convert time into C style timeval."""
-    frac, whole = math.modf(seconds_since_epoch)
-    microseconds = math.floor(frac * 1000000)
-    seconds = math.floor(whole)
-    return seconds, microseconds
 
 
 # Now comes all the structs we need to parse the infomation coming
@@ -185,208 +128,6 @@ class MSLLHookStruct(ctypes.Structure):
         ("extrainfo", ctypes.c_ulong),
     ]
 
-
-class XinputGamepad(ctypes.Structure):
-    """Describes the current state of the Xbox 360 Controller.
-
-    For full details see Microsoft's documentation:
-
-    https://msdn.microsoft.com/en-us/library/windows/desktop/
-    microsoft.directx_sdk.reference.xinput_gamepad%28v=vs.85%29.aspx
-
-    """
-
-    # pylint: disable=too-few-public-methods
-    _fields_ = [
-        ("buttons", ctypes.c_ushort),  # wButtons
-        ("left_trigger", ctypes.c_ubyte),  # bLeftTrigger
-        ("right_trigger", ctypes.c_ubyte),  # bLeftTrigger
-        ("l_thumb_x", ctypes.c_short),  # sThumbLX
-        ("l_thumb_y", ctypes.c_short),  # sThumbLY
-        ("r_thumb_x", ctypes.c_short),  # sThumbRx
-        ("r_thumb_y", ctypes.c_short),  # sThumbRy
-    ]
-
-
-class XinputState(ctypes.Structure):
-    """Represents the state of a controller.
-
-    For full details see Microsoft's documentation:
-
-    https://msdn.microsoft.com/en-us/library/windows/desktop/
-    microsoft.directx_sdk.reference.xinput_state%28v=vs.85%29.aspx
-
-    """
-
-    # pylint: disable=too-few-public-methods
-    _fields_ = [
-        ("packet_number", ctypes.c_ulong),  # dwPacketNumber
-        ("gamepad", XinputGamepad),  # Gamepad
-    ]
-
-
-class XinputVibration(ctypes.Structure):
-    """Specifies motor speed levels for the vibration function of a
-    controller.
-
-    For full details see Microsoft's documentation:
-
-    https://msdn.microsoft.com/en-us/library/windows/desktop/
-    microsoft.directx_sdk.reference.xinput_vibration%28v=vs.85%29.aspx
-
-    """
-
-    # pylint: disable=too-few-public-methods
-    _fields_ = [
-        ("wLeftMotorSpeed", ctypes.c_ushort),
-        ("wRightMotorSpeed", ctypes.c_ushort),
-    ]
-
-
-if sys.version_info.major == 2:
-    # pylint: disable=redefined-builtin
-    class PermissionError(IOError):
-        """Raised when trying to run an operation without the adequate access
-        rights - for example filesystem permissions. Corresponds to errno
-        EACCES and EPERM."""
-
-
-class InputEvent(object):  # pylint: disable=useless-object-inheritance
-    """A user event."""
-
-    # pylint: disable=too-few-public-methods
-    def __init__(self, device, event_info):
-        self.device = device
-        self.timestamp = event_info["timestamp"]
-        self.code = event_info["code"]
-        self.state = event_info["state"]
-        self.ev_type = event_info["ev_type"]
-
-
-class BaseListener(object):  # pylint: disable=useless-object-inheritance
-    """Loosely emulate Evdev keyboard behaviour on other platforms.
-    Listen (hook in Windows terminology) for key events then buffer
-    them in a pipe.
-    """
-
-    def __init__(self, pipe, events=None, codes=None):
-        self.pipe = pipe
-        self.events = events if events else []
-        self.codes = codes if codes else None
-        self.app = None
-        self.timeval = None
-        self.type_codes = dict(((value, key) for key, value in EVENT_TYPES))
-
-        self.install_handle_input()
-
-    def install_handle_input(self):
-        """Install the input handler."""
-        pass
-
-    def uninstall_handle_input(self):
-        """Un-install the input handler."""
-        pass
-
-    def __del__(self):
-        """Clean up when deleted."""
-        self.uninstall_handle_input()
-
-    @staticmethod
-    def get_timeval():
-        """Get the time in seconds and microseconds."""
-        return convert_timeval(time.time())
-
-    def update_timeval(self):
-        """Update the timeval with the current time."""
-        self.timeval = self.get_timeval()
-
-    def create_event_object(self, event_type, code, value, timeval=None):
-        """Create an evdev style structure."""
-        if not timeval:
-            self.update_timeval()
-            timeval = self.timeval
-        try:
-            event_code = self.type_codes[event_type]
-        except KeyError:
-            raise UnknownEventType(
-                "We don't know what kind of event a %s is." % event_type
-            )
-
-        event = struct.pack(
-            EVENT_FORMAT, timeval[0], timeval[1], event_code, code, value
-        )
-        return event
-
-    def write_to_pipe(self, event_list):
-        """Send event back to the mouse object."""
-        self.pipe.send_bytes(b"".join(event_list))
-
-    def emulate_wheel(self, data, direction, timeval):
-        """Emulate rel values for the mouse wheel.
-
-        In evdev, a single click forwards of the mouse wheel is 1 and
-        a click back is -1. Windows uses 120 and -120. We floor divide
-        the Windows number by 120. This is fine for the digital scroll
-        wheels found on the vast majority of mice. It also works on
-        the analogue ball on the top of the Apple mouse.
-
-        What do the analogue scroll wheels found on 200 quid high end
-        gaming mice do? If the lowest unit is 120 then we are okay. If
-        they report changes of less than 120 units Windows, then this
-        might be an unacceptable loss of precision. Needless to say, I
-        don't have such a mouse to test one way or the other.
-
-        """
-        if direction == "x":
-            code = 0x06
-        elif direction == "z":
-            # Not enitely sure if this exists
-            code = 0x07
-        else:
-            code = 0x08
-
-        if WIN:
-            data = data // 120
-
-        return self.create_event_object("Relative", code, data, timeval)
-
-    def emulate_rel(self, key_code, value, timeval):
-        """Emulate the relative changes of the mouse cursor."""
-        return self.create_event_object("Relative", key_code, value, timeval)
-
-    def emulate_press(self, key_code, scan_code, value, timeval):
-        """Emulate a button press.
-
-        Currently supports 5 buttons.
-
-        The Microsoft documentation does not define what happens with
-        a mouse with more than five buttons, and I don't have such a
-        mouse.
-
-        From reading the Linux sources, I guess evdev can support up
-        to 255 buttons.
-
-        Therefore, I guess we could support more buttons quite easily,
-        if we had any useful hardware.
-        """
-        scan_event = self.create_event_object("Misc", 0x04, scan_code, timeval)
-        key_event = self.create_event_object("Key", key_code, value, timeval)
-        return scan_event, key_event
-
-    def emulate_repeat(self, value, timeval):
-        """The repeat press of a key/mouse button, e.g. double click."""
-        repeat_event = self.create_event_object("Repeat", 2, value, timeval)
-        return repeat_event
-
-    def sync_marker(self, timeval):
-        """Separate groups of events."""
-        return self.create_event_object("Sync", 0, 0, timeval)
-
-    def emulate_abs(self, x_val, y_val, timeval):
-        """Emulate the absolute co-ordinates of the mouse cursor."""
-        x_event = self.create_event_object("Absolute", 0x00, x_val, timeval)
-        y_event = self.create_event_object("Absolute", 0x01, y_val, timeval)
-        return x_event, y_event
 
 
 class WindowsKeyboardListener(BaseListener):
@@ -1122,183 +863,6 @@ def mac_keyboard_process(pipe):
     keyboard = MacKeyboardListener(pipe)
 
 
-class InputDevice(object):  # pylint: disable=useless-object-inheritance
-    """A user input device."""
-
-    # pylint: disable=too-many-instance-attributes
-    def __init__(self, manager, device_path=None, char_path_override=None, read_size=1):
-        self.read_size = read_size
-        self.manager = manager
-        self.__pipe = None
-        self._listener = None
-        self.leds = None
-        if device_path:
-            self._device_path = device_path
-        else:
-            self._set_device_path()
-        # We should by now have a device_path
-
-        try:
-            if not self._device_path:
-                raise NoDevicePath
-        except AttributeError:
-            raise NoDevicePath
-
-        self.protocol, _, self.device_type = self._get_path_infomation()
-        if char_path_override:
-            self._character_device_path = char_path_override
-        else:
-            self._character_device_path = os.path.realpath(self._device_path)
-
-        self._character_file = None
-
-        self._evdev = False
-        self._set_evdev_state()
-
-        self.name = "Unknown Device"
-        self._set_name()
-
-    def _set_device_path(self):
-        """Set the device path, overridden on the MAC and Windows."""
-        pass
-
-    def _set_evdev_state(self):
-        """Set whether the device is a real evdev device."""
-        if NIX:
-            self._evdev = True
-
-    def _set_name(self):
-        if NIX:
-            with open(
-                "/sys/class/input/%s/device/name" % self.get_char_name()
-            ) as name_file:
-                self.name = name_file.read().strip()
-            self.leds = []
-
-    def _get_path_infomation(self):
-        """Get useful infomation from the device path."""
-        long_identifier = self._device_path.split("/")[4]
-        protocol, remainder = long_identifier.split("-", 1)
-        identifier, _, device_type = remainder.rsplit("-", 2)
-        return (protocol, identifier, device_type)
-
-    def get_char_name(self):
-        """Get short version of char device name."""
-        return self._character_device_path.split("/")[-1]
-
-    def get_char_device_path(self):
-        """Get the char device path."""
-        return self._character_device_path
-
-    def __str__(self):
-        try:
-            return self.name
-        except AttributeError:
-            return "Unknown Device"
-
-    def __repr__(self):
-        return '%s.%s("%s")' % (
-            self.__module__,
-            self.__class__.__name__,
-            self._device_path,
-        )
-
-    @property
-    def _character_device(self):
-        if not self._character_file:
-            if WIN:
-                self._character_file = io.BytesIO()
-                return self._character_file
-            try:
-                self._character_file = io.open(self._character_device_path, "rb")
-            except PermissionError:
-                # Python 3
-                raise PermissionError(PERMISSIONS_ERROR_TEXT)
-            except IOError as err:
-                # Python 2
-                if err.errno == 13:
-                    raise PermissionError(PERMISSIONS_ERROR_TEXT)
-                else:
-                    raise
-
-        return self._character_file
-
-    def __iter__(self):
-        while True:
-            event = self._do_iter()
-            if event:
-                yield event
-
-    def _get_data(self, read_size):
-        """Get data from the character device."""
-        return self._character_device.read(read_size)
-
-    @staticmethod
-    def _get_target_function():
-        """Get the correct target function. This is only used by Windows
-        subclasses."""
-        return False
-
-    def _get_total_read_size(self):
-        """How much event data to process at once."""
-        if self.read_size:
-            read_size = EVENT_SIZE * self.read_size
-        else:
-            read_size = EVENT_SIZE
-        return read_size
-
-    def _do_iter(self):
-        read_size = self._get_total_read_size()
-        data = self._get_data(read_size)
-        if not data:
-            return None
-        evdev_objects = iter_unpack(data)
-        events = [self._make_event(*event) for event in evdev_objects]
-        return events
-
-    # pylint: disable=too-many-arguments
-    def _make_event(self, tv_sec, tv_usec, ev_type, code, value):
-        """Create a friendly Python object from an evdev style event."""
-        event_type = self.manager.get_event_type(ev_type)
-        eventinfo = {
-            "ev_type": event_type,
-            "state": value,
-            "timestamp": tv_sec + (tv_usec / 1000000),
-            "code": self.manager.get_event_string(event_type, code),
-        }
-
-        return InputEvent(self, eventinfo)
-
-    def read(self):
-        """Read the next input event."""
-        return next(iter(self))
-
-    @property
-    def _pipe(self):
-        """On Windows we use a pipe to emulate a Linux style character
-        buffer."""
-        if self._evdev:
-            return None
-
-        if not self.__pipe:
-            target_function = self._get_target_function()
-            if not target_function:
-                return None
-
-            self.__pipe, child_conn = Pipe(duplex=False)
-            self._listener = Process(
-                target=target_function, args=(child_conn,), daemon=True
-            )
-            self._listener.start()
-        return self.__pipe
-
-    def __del__(self):
-        if "WIN" in globals() or "MAC" in globals():
-            if WIN or MAC:
-                if self.__pipe:
-                    self._listener.terminate()
-
-
 class Keyboard(InputDevice):
     """A keyboard or other key-like device.
 
@@ -1381,347 +945,6 @@ class MightyMouse(Mouse):
         """Get the correct target function."""
         return quartz_mouse_process
 
-
-def delay_and_stop(duration, dll, device_number):
-    """Stop vibration aka force feedback aka rumble on
-    Windows after duration miliseconds."""
-    xinput = getattr(ctypes.windll, dll)
-    time.sleep(duration / 1000)
-    xinput_set_state = xinput.XInputSetState
-    xinput_set_state.argtypes = [ctypes.c_uint, ctypes.POINTER(XinputVibration)]
-    xinput_set_state.restype = ctypes.c_uint
-    vibration = XinputVibration(0, 0)
-    xinput_set_state(device_number, ctypes.byref(vibration))
-
-
-# I made this GamePad class before Mouse and Keyboard above, and have
-# learned a lot about Windows in the process.  This can probably be
-# simplified massively and made to match Mouse and Keyboard more.
-
-
-class GamePad(InputDevice):
-    """A gamepad or other joystick-like device."""
-
-    def __init__(self, manager, device_path, char_path_override=None):
-        super(GamePad, self).__init__(manager, device_path, char_path_override)
-        self._write_file = None
-        self.__device_number = None
-        if WIN:
-            if "Microsoft_Corporation_Controller" in self._device_path:
-                self.name = "Microsoft X-Box 360 pad"
-                identifier = self._get_path_infomation()[1]
-                self.__device_number = int(identifier.split("_")[-1])
-                self.__received_packets = 0
-                self.__missed_packets = 0
-                self.__last_state = self.__read_device()
-        if NIX:
-            self._number_xpad()
-
-    def _number_xpad(self):
-        """Get the number of the joystick."""
-        js_path = self._device_path.replace("-event", "")
-        js_chardev = os.path.realpath(js_path)
-        try:
-            number_text = js_chardev.split("js")[1]
-        except IndexError:
-            return
-        try:
-            number = int(number_text)
-        except ValueError:
-            return
-        self.__device_number = number
-
-    def get_number(self):
-        """Return the joystick number of the gamepad."""
-        return self.__device_number
-
-    def __iter__(self):
-        while True:
-            if WIN:
-                self.__check_state()
-            event = self._do_iter()
-            if event:
-                yield event
-
-    def __check_state(self):
-        """On Windows, check the state and fill the event character device."""
-        state = self.__read_device()
-        if not state:
-            raise UnpluggedError("Gamepad %d is not connected" % self.__device_number)
-        if state.packet_number != self.__last_state.packet_number:
-            # state has changed, handle the change
-            self.__handle_changed_state(state)
-            self.__last_state = state
-
-    @staticmethod
-    def __get_timeval():
-        """Get the time and make it into C style timeval."""
-        return convert_timeval(time.time())
-
-    def create_event_object(self, event_type, code, value, timeval=None):
-        """Create an evdev style object."""
-        if not timeval:
-            timeval = self.__get_timeval()
-        try:
-            event_code = self.manager.codes["type_codes"][event_type]
-        except KeyError:
-            raise UnknownEventType(
-                "We don't know what kind of event a %s is." % event_type
-            )
-        event = struct.pack(
-            EVENT_FORMAT, timeval[0], timeval[1], event_code, code, value
-        )
-        return event
-
-    def __write_to_character_device(self, event_list, timeval=None):
-        """Emulate the Linux character device on other platforms such as
-        Windows."""
-        # Remember the position of the stream
-        pos = self._character_device.tell()
-        # Go to the end of the stream
-        self._character_device.seek(0, 2)
-        # Write the new data to the end
-        for event in event_list:
-            self._character_device.write(event)
-        # Add a sync marker
-        sync = self.create_event_object("Sync", 0, 0, timeval)
-        self._character_device.write(sync)
-        # Put the stream back to its original position
-        self._character_device.seek(pos)
-
-    def __handle_changed_state(self, state):
-        """
-        we need to pack a struct with the following five numbers:
-        tv_sec, tv_usec, ev_type, code, value
-
-        then write it using __write_to_character_device
-
-        seconds, mircroseconds, ev_type, code, value
-        time we just use now
-        ev_type we look up
-        code we look up
-        value is 0 or 1 for the buttons
-        axis value is maybe the same as Linux? Hope so!
-        """
-        timeval = self.__get_timeval()
-        events = self.__get_button_events(state, timeval)
-        events.extend(self.__get_axis_events(state, timeval))
-        if events:
-            self.__write_to_character_device(events, timeval)
-
-    def __map_button(self, button):
-        """Get the linux xpad code from the Windows xinput code."""
-        _, start_code, start_value = button
-        value = start_value
-        ev_type = "Key"
-        code = self.manager.codes["xpad"][start_code]
-        if 1 <= start_code <= 4:
-            ev_type = "Absolute"
-        if start_code == 1 and start_value == 1:
-            value = -1
-        elif start_code == 3 and start_value == 1:
-            value = -1
-        return code, value, ev_type
-
-    def __map_axis(self, axis):
-        """Get the linux xpad code from the Windows xinput code."""
-        start_code, start_value = axis
-        value = start_value
-        code = self.manager.codes["xpad"][start_code]
-        return code, value
-
-    def __get_button_events(self, state, timeval=None):
-        """Get the button events from xinput."""
-        changed_buttons = self.__detect_button_events(state)
-        events = self.__emulate_buttons(changed_buttons, timeval)
-        return events
-
-    def __get_axis_events(self, state, timeval=None):
-        """Get the stick events from xinput."""
-        axis_changes = self.__detect_axis_events(state)
-        events = self.__emulate_axis(axis_changes, timeval)
-        return events
-
-    def __emulate_axis(self, axis_changes, timeval=None):
-        """Make the axis events use the Linux style format."""
-        events = []
-        for axis in axis_changes:
-            code, value = self.__map_axis(axis)
-            event = self.create_event_object("Absolute", code, value, timeval=timeval)
-            events.append(event)
-        return events
-
-    def __emulate_buttons(self, changed_buttons, timeval=None):
-        """Make the button events use the Linux style format."""
-        events = []
-        for button in changed_buttons:
-            code, value, ev_type = self.__map_button(button)
-            event = self.create_event_object(ev_type, code, value, timeval=timeval)
-            events.append(event)
-        return events
-
-    @staticmethod
-    def __gen_bit_values(number):
-        """
-        Return a zero or one for each bit of a numeric value up to the most
-        significant 1 bit, beginning with the least significant bit.
-        """
-        number = int(number)
-        while number:
-            yield number & 0x1
-            number >>= 1
-
-    def __get_bit_values(self, number, size=32):
-        """Get bit values as a list for a given number
-
-        >>> get_bit_values(1) == [0]*31 + [1]
-        True
-
-        >>> get_bit_values(0xDEADBEEF)
-        [1L, 1L, 0L, 1L, 1L, 1L, 1L,
-        0L, 1L, 0L, 1L, 0L, 1L, 1L, 0L, 1L, 1L, 0L, 1L, 1L, 1L, 1L,
-        1L, 0L, 1L, 1L, 1L, 0L, 1L, 1L, 1L, 1L]
-
-        You may override the default word size of 32-bits to match your actual
-        application.
-        >>> get_bit_values(0x3, 2)
-        [1L, 1L]
-
-        >>> get_bit_values(0x3, 4)
-        [0L, 0L, 1L, 1L]
-
-        """
-        res = list(self.__gen_bit_values(number))
-        res.reverse()
-        # 0-pad the most significant bit
-        res = [0] * (size - len(res)) + res
-        return res
-
-    def __detect_button_events(self, state):
-        changed = state.gamepad.buttons ^ self.__last_state.gamepad.buttons
-        changed = self.__get_bit_values(changed, 16)
-        buttons_state = self.__get_bit_values(state.gamepad.buttons, 16)
-        changed.reverse()
-        buttons_state.reverse()
-        button_numbers = count(1)
-        changed_buttons = list(
-            filter(itemgetter(0), list(zip(changed, button_numbers, buttons_state)))
-        )
-        # returns for example [(1,15,1)] type, code, value?
-        return changed_buttons
-
-    def __detect_axis_events(self, state):
-        # axis fields are everything but the buttons
-        # pylint: disable=protected-access
-        # Attribute name _fields_ is special name set by ctypes
-        axis_fields = dict(XinputGamepad._fields_)
-        axis_fields.pop("buttons")
-        changed_axes = []
-
-        # Ax_type might be useful when we support high-level deadzone
-        # methods.
-        # pylint: disable=unused-variable
-        for axis, ax_type in list(axis_fields.items()):
-            old_val = getattr(self.__last_state.gamepad, axis)
-            new_val = getattr(state.gamepad, axis)
-            if old_val != new_val:
-                changed_axes.append((axis, new_val))
-        return changed_axes
-
-    def __read_device(self):
-        """Read the state of the gamepad."""
-        state = XinputState()
-        res = self.manager.xinput.XInputGetState(
-            self.__device_number, ctypes.byref(state)
-        )
-        if res == XINPUT_ERROR_SUCCESS:
-            return state
-        if res != XINPUT_ERROR_DEVICE_NOT_CONNECTED:
-            raise RuntimeError(
-                "Unknown error %d attempting to get state of device %d"
-                % (res, self.__device_number)
-            )
-        # else (device is not connected)
-        return None
-
-    @property
-    def _write_device(self):
-        if not self._write_file:
-            if not NIX:
-                return None
-            try:
-                self._write_file = io.open(self._character_device_path, "wb")
-            except PermissionError:
-                # Python 3
-                raise PermissionError(PERMISSIONS_ERROR_TEXT)
-            except IOError as err:
-                # Python 2
-                if err.errno == 13:
-                    raise PermissionError(PERMISSIONS_ERROR_TEXT)
-                else:
-                    raise
-
-        return self._write_file
-
-    def _start_vibration_win(self, left_motor, right_motor):
-        """Start the vibration, which will run until stopped."""
-        xinput_set_state = self.manager.xinput.XInputSetState
-        xinput_set_state.argtypes = [ctypes.c_uint, ctypes.POINTER(XinputVibration)]
-        xinput_set_state.restype = ctypes.c_uint
-        vibration = XinputVibration(int(left_motor * 65535), int(right_motor * 65535))
-        xinput_set_state(self.__device_number, ctypes.byref(vibration))
-
-    def _stop_vibration_win(self):
-        """Stop the vibration."""
-        xinput_set_state = self.manager.xinput.XInputSetState
-        xinput_set_state.argtypes = [ctypes.c_uint, ctypes.POINTER(XinputVibration)]
-        xinput_set_state.restype = ctypes.c_uint
-        stop_vibration = ctypes.byref(XinputVibration(0, 0))
-        xinput_set_state(self.__device_number, stop_vibration)
-
-    def _set_vibration_win(self, left_motor, right_motor, duration):
-        """Control the motors on Windows."""
-        self._start_vibration_win(left_motor, right_motor)
-        stop_process = Process(
-            target=delay_and_stop,
-            args=(duration, self.manager.xinput_dll, self.__device_number),
-        )
-        stop_process.start()
-
-    def __get_vibration_code(self, left_motor, right_motor, duration):
-        """This is some crazy voodoo, if you can simplify it, please do."""
-        inner_event = struct.pack(
-            "2h6x2h2x2H28x",
-            0x50,
-            -1,
-            duration,
-            0,
-            int(left_motor * 65535),
-            int(right_motor * 65535),
-        )
-        buf_conts = ioctl(self._write_device, 1076905344, inner_event)
-        return int(codecs.encode(buf_conts[1:3], "hex"), 16)
-
-    def _set_vibration_nix(self, left_motor, right_motor, duration):
-        """Control the motors on Linux.
-        Duration is in miliseconds."""
-        code = self.__get_vibration_code(left_motor, right_motor, duration)
-        secs, msecs = convert_timeval(time.time())
-        outer_event = struct.pack(EVENT_FORMAT, secs, msecs, 0x15, code, 1)
-        self._write_device.write(outer_event)
-        self._write_device.flush()
-
-    def set_vibration(self, left_motor, right_motor, duration):
-        """Control the speed of both motors seperately or together.
-        left_motor and right_motor arguments require a number between
-        0 (off) and 1 (full).
-        duration is miliseconds, e.g. 1000 for a second."""
-        if WIN:
-            self._set_vibration_win(left_motor, right_motor, duration)
-        elif NIX:
-            self._set_vibration_nix(left_motor, right_motor, duration)
-        else:
-            raise NotImplementedError
 
 
 class OtherDevice(InputDevice):
@@ -2177,237 +1400,3 @@ class DeviceManager(object):  # pylint: disable=useless-object-inheritance
             self.gamepads.append(gpad)
 
 
-SPIN_UP_MOTOR = (
-    "00000",
-    "00001",
-    "00011",
-    "00111",
-    "01111",
-    "11111",
-    "01111",
-    "00011",
-    "00001",
-    "00000",
-    "00001",
-    "00011",
-    "00111",
-    "01111",
-    "11111",
-    "00000",
-    "11111",
-    "00000",
-    "11111",
-    "00000",
-)
-
-
-class MicroBitPad(GamePad):
-    """A BBC Micro:bit flashed with bitio."""
-
-    def __init__(self, manager, device_path=None, char_path_override=None):
-        if not device_path:
-            device_path = "/dev/input/by-id/dialup-BBC_MicroBit-event-joystick"
-            if not char_path_override:
-                char_path_override = "/dev/input/microbit0"
-
-        super(MicroBitPad, self).__init__(manager, device_path, char_path_override)
-
-        # pylint: disable=no-member,import-error
-        import microbit
-
-        self.microbit = microbit
-        self.default_image = microbit.Image("00500:00500:00500:00500:00500")
-        self._setup_rumble()
-        self.set_display()
-
-    def set_display(self, index=None):
-        """Show an image on the display."""
-        # pylint: disable=no-member
-        if index:
-            image = self.microbit.Image.STD_IMAGES[index]
-        else:
-            image = self.default_image
-        self.microbit.display.show(image)
-
-    def _setup_rumble(self):
-        """Setup the three animations which simulate a rumble."""
-        self.left_rumble = self._get_ready_to("99500")
-        self.right_rumble = self._get_ready_to("00599")
-        self.double_rumble = self._get_ready_to("99599")
-
-    def _set_name(self):
-        self.name = "BBC microbit Gamepad"
-
-    def _set_evdev_state(self):
-        self._evdev = False
-
-    @staticmethod
-    def _get_target_function():
-        return microbit_process
-
-    def _get_data(self, read_size):
-        """Get data from the character device."""
-        return self._pipe.recv_bytes()
-
-    def _get_ready_to(self, rumble):
-        """Watch us wreck the mike!
-        PSYCHE!"""
-        # pylint: disable=no-member
-        return [
-            self.microbit.Image(
-                ":".join([rumble if char == "1" else "00500" for char in code])
-            )
-            for code in SPIN_UP_MOTOR
-        ]
-
-    def _full_speed_rumble(self, images, duration):
-        """Simulate the motors running at full."""
-        while duration > 0:
-            self.microbit.display.show(images[0])  # pylint: disable=no-member
-            time.sleep(0.04)
-            self.microbit.display.show(images[1])  # pylint: disable=no-member
-            time.sleep(0.04)
-            duration -= 0.08
-
-    def _spin_up(self, images, duration):
-        """Simulate the motors getting warmed up."""
-        total = 0
-        # pylint: disable=no-member
-
-        for image in images:
-            self.microbit.display.show(image)
-            time.sleep(0.05)
-            total += 0.05
-            if total >= duration:
-                return
-        remaining = duration - total
-        self._full_speed_rumble(images[-2:], remaining)
-        self.set_display()
-
-    def set_vibration(self, left_motor, right_motor, duration):
-        """Control the speed of both motors seperately or together.
-        left_motor and right_motor arguments require a number:
-        0 (off) or 1 (full).
-        duration is miliseconds, e.g. 1000 for a second."""
-        if left_motor and right_motor:
-            return self._spin_up(self.double_rumble, duration / 1000)
-        if left_motor:
-            return self._spin_up(self.left_rumble, duration / 1000)
-        if right_motor:
-            return self._spin_up(self.right_rumble, duration / 1000)
-        return -1
-
-
-def microbit_process(pipe):
-    """Simple subprocess for reading mouse events on the microbit."""
-    gamepad_listener = MicroBitListener(pipe)
-    gamepad_listener.listen()
-
-
-class MicroBitListener(BaseListener):
-    """Tracks the current state and sends changes to the MicroBitPad
-    device class."""
-
-    def __init__(self, pipe):
-        super(MicroBitListener, self).__init__(pipe)
-        self.active = True
-        self.events = []
-        self.state = set(
-            (
-                ("Absolute", 0x10, 0),
-                ("Absolute", 0x11, 0),
-                ("Key", 0x130, 0),
-                ("Key", 0x131, 0),
-                ("Key", 0x13A, 0),
-                ("Key", 0x133, 0),
-                ("Key", 0x134, 0),
-            )
-        )
-        self.dpad = True
-        self.sensitivity = 300
-        # pylint: disable=import-error
-        import microbit
-
-        self.microbit = microbit
-
-    def listen(self):
-        """Listen while the device is active."""
-        while self.active:
-            self.handle_input()
-
-    def uninstall_handle_input(self):
-        """Stop listing when active is false."""
-        self.active = False
-
-    def handle_new_events(self, events):
-        """Add each new events to the event queue."""
-        for event in events:
-            self.events.append(
-                self.create_event_object(event[0], event[1], int(event[2]))
-            )
-
-    def handle_abs(self):
-        """Gets the state as the raw abolute numbers."""
-        # pylint: disable=no-member
-        x_raw = self.microbit.accelerometer.get_x()
-        y_raw = self.microbit.accelerometer.get_y()
-        x_abs = ("Absolute", 0x00, x_raw)
-        y_abs = ("Absolute", 0x01, y_raw)
-        return x_abs, y_abs
-
-    def handle_dpad(self):
-        """Gets the state of the virtual dpad."""
-        # pylint: disable=no-member
-        x_raw = self.microbit.accelerometer.get_x()
-        y_raw = self.microbit.accelerometer.get_y()
-        minus_sens = self.sensitivity * -1
-        if x_raw < minus_sens:
-            x_state = ("Absolute", 0x10, -1)
-        elif x_raw > self.sensitivity:
-            x_state = ("Absolute", 0x10, 1)
-        else:
-            x_state = ("Absolute", 0x10, 0)
-
-        if y_raw < minus_sens:
-            y_state = ("Absolute", 0x11, -1)
-        elif y_raw > self.sensitivity:
-            y_state = ("Absolute", 0x11, 1)
-        else:
-            y_state = ("Absolute", 0x11, 1)
-
-        return x_state, y_state
-
-    def check_state(self):
-        """Tracks differences in the device state."""
-        if self.dpad:
-            x_state, y_state = self.handle_dpad()
-        else:
-            x_state, y_state = self.handle_abs()
-
-        # pylint: disable=no-member
-        new_state = set(
-            (
-                x_state,
-                y_state,
-                ("Key", 0x130, int(self.microbit.button_a.is_pressed())),
-                ("Key", 0x131, int(self.microbit.button_b.is_pressed())),
-                ("Key", 0x13A, int(self.microbit.pin0.is_touched())),
-                ("Key", 0x133, int(self.microbit.pin1.is_touched())),
-                ("Key", 0x134, int(self.microbit.pin2.is_touched())),
-            )
-        )
-        events = new_state - self.state
-        self.state = new_state
-        return events
-
-    def handle_input(self):
-        """Sends differences in the device state to the MicroBitPad
-        as events."""
-        difference = self.check_state()
-        if not difference:
-            return
-        self.events = []
-        self.handle_new_events(difference)
-        self.update_timeval()
-        self.events.append(self.sync_marker(self.timeval))
-        self.write_to_pipe(self.events)
