+++ tests/test_listeners.py
@@ -4,6 +4,11 @@ from unittest import TestCase
 
 import inputs
 
+import inputs.platforms
+from inputs.platforms.baselistener import BaseListener
+import inputs.platforms.baselistener
+from inputs.platforms.c import iter_unpack
+import inputs.platforms.system
 from tests.constants import mock, PYTHON
 if PYTHON == 3:
     mock._magics.add('__round__')
@@ -28,13 +33,13 @@ class BaseListenerTestCase(TestCase):
     def test_init(self):
         """The listener has type_codes."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         self.assertEqual(len(listener.type_codes), 14)
 
     def test_get_timeval(self):
         """Gives seconds and microseconds."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         seconds, microseconds = listener.get_timeval()
         self.assertTrue(seconds > 0)
         self.assertTrue(microseconds > 0)
@@ -42,7 +47,7 @@ class BaseListenerTestCase(TestCase):
     def test_set_timeval(self):
         """Sets the cached timeval."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         # We start with no timeval
         self.assertIsNone(listener.timeval)
@@ -56,32 +61,32 @@ class BaseListenerTestCase(TestCase):
     def test_create_key_event_object(self):
         """It should create an evdev object."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.create_event_object("Key", 30, 1, (100, 0))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 0, 1, 30, 1))
 
     def test_create_mouse_event_object(self):
         """It also should create an evdev object."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.create_event_object("Absolute", 0, 285, (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 3, 0, 285))
 
     def test_create_banana_event_object(self):
         """It should raise an exception."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         with self.assertRaises(inputs.UnknownEventType):
             listener.create_event_object("Banana", 0, 285, (100, 1))
 
     def test_create_ev_wo_timeval(self):
         """It should create an evdev object."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.create_event_object("Key", 30, 1)
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertTrue(event_info[0] > 0)
         self.assertTrue(event_info[1] > 0)
         self.assertEqual(event_info[2:], (1, 30, 1))
@@ -89,7 +94,7 @@ class BaseListenerTestCase(TestCase):
     def test_write_to_pipe(self):
         """Subprocess sends data back to the class in the mainprocess."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         listener.write_to_pipe([b'Green Eggs', b' and ', b'Ham'])
         send_bytes_call = pipe.method_calls[0]
         method_name = send_bytes_call[0]
@@ -100,83 +105,83 @@ class BaseListenerTestCase(TestCase):
     def test_emulate_wheel_x(self):
         """Returns an event list for the x mouse wheel turn."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.emulate_wheel(20, 'x', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 6, 20))
 
         eventlist = listener.emulate_wheel(-20, 'x', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 6, -20))
 
     def test_emulate_wheel_y(self):
         """Returns an event list for the y mouse wheel turn."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.emulate_wheel(20, 'y', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 8, 20))
 
         eventlist = listener.emulate_wheel(-20, 'y', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 8, -20))
 
     def test_emulate_wheel_z(self):
         """Returns an event list for the z mouse wheel turn."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.emulate_wheel(20, 'z', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 7, 20))
 
         eventlist = listener.emulate_wheel(-20, 'z', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 7, -20))
 
     def test_emulate_wheel_win(self):
         """Returns an event list for the mouse wheel turn on Windows."""
-        inputs.WIN = True
+        inputs.platforms.baselistener.WIN = True
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
         eventlist = listener.emulate_wheel(240, 'x', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 6, 2))
 
         eventlist = listener.emulate_wheel(-240, 'x', (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 6, -2))
-        inputs.WIN = False
+        inputs.platforms.baselistener.WIN = False
 
     def test_emulate_rel(self):
         """Returns an event list for relative mouse movement."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         eventlist = listener.emulate_rel(0, 1, (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 0, 1))
 
         eventlist = listener.emulate_rel(0, -5, (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 0, -5))
 
         eventlist = listener.emulate_rel(1, 44, (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 1, 44))
 
         eventlist = listener.emulate_rel(1, -10, (100, 1))
-        event_info = next(inputs.iter_unpack(eventlist))
+        event_info = next(iter_unpack(eventlist))
         self.assertEqual(event_info, (100, 1, 2, 1, -10))
 
     def test_emulate_press_down(self):
         """Returns an event list for button."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         scan_list, button_list = listener.emulate_press(
             272, 589825, 1, (100, 1))
-        scan_info = next(inputs.iter_unpack(scan_list))
-        button_info = next(inputs.iter_unpack(button_list))
+        scan_info = next(iter_unpack(scan_list))
+        button_info = next(iter_unpack(button_list))
 
         self.assertEqual(scan_info, (100, 1, 4, 4, 589825))
         self.assertEqual(button_info, (100, 1, 1, 272, 1))
@@ -184,12 +189,12 @@ class BaseListenerTestCase(TestCase):
     def test_emulate_press_up(self):
         """Returns an event list for button."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         scan_list, button_list = listener.emulate_press(
             272, 589825, 0, (100, 1))
-        scan_info = next(inputs.iter_unpack(scan_list))
-        button_info = next(inputs.iter_unpack(button_list))
+        scan_info = next(iter_unpack(scan_list))
+        button_info = next(iter_unpack(button_list))
 
         self.assertEqual(scan_info, (100, 1, 4, 4, 589825))
         self.assertEqual(button_info, (100, 1, 1, 272, 0))
@@ -197,42 +202,42 @@ class BaseListenerTestCase(TestCase):
     def test_emulate_repeat(self):
         """Returns a repeat event, e.g. doubleclick, triple click."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         repeat_list = listener.emulate_repeat(1, (100, 1))
-        repeat_info = next(inputs.iter_unpack(repeat_list))
+        repeat_info = next(iter_unpack(repeat_list))
         self.assertEqual(repeat_info, (100, 1, 20, 2, 1))
 
         repeat_list = listener.emulate_repeat(2, (100, 1))
-        repeat_info = next(inputs.iter_unpack(repeat_list))
+        repeat_info = next(iter_unpack(repeat_list))
         self.assertEqual(repeat_info, (100, 1, 20, 2, 2))
 
         repeat_list = listener.emulate_repeat(3, (100, 1))
-        repeat_info = next(inputs.iter_unpack(repeat_list))
+        repeat_info = next(iter_unpack(repeat_list))
         self.assertEqual(repeat_info, (100, 1, 20, 2, 3))
 
     def test_sync_marker(self):
         """Returns a sync marker."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         sync_list = listener.sync_marker((100, 1))
-        sync_info = next(inputs.iter_unpack(sync_list))
+        sync_info = next(iter_unpack(sync_list))
         self.assertEqual(sync_info, (100, 1, 0, 0, 0))
 
         sync_list = listener.sync_marker((200, 2))
-        sync_info = next(inputs.iter_unpack(sync_list))
+        sync_info = next(iter_unpack(sync_list))
         self.assertEqual(sync_info, (200, 2, 0, 0, 0))
 
     def test_emulate_abs(self):
         """Returns absolute mouse event."""
         pipe = mock.MagicMock()
-        listener = inputs.BaseListener(pipe)
+        listener = BaseListener(pipe)
 
         x_list, y_list = listener.emulate_abs(1324, 246, (100, 1))
-        x_info = next(inputs.iter_unpack(x_list))
+        x_info = next(iter_unpack(x_list))
         self.assertEqual(x_info, (100, 1, 3, 0, 1324))
-        y_info = next(inputs.iter_unpack(y_list))
+        y_info = next(iter_unpack(y_list))
         self.assertEqual(y_info, (100, 1, 3, 1, 246))
 
 
@@ -306,13 +311,13 @@ class QuartzMouseBaseListenerTestCase(TestCase):
         # Now there are three events
         self.assertEqual(len(listener.events), 3)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (4, 4, 589826))
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (1, 273, 1))
-        third_event = next(inputs.iter_unpack(
+        third_event = next(iter_unpack(
             listener.events[2]))
         self.assertEqual(third_event[2:], (20, 2, 1))
 
@@ -345,13 +350,13 @@ class QuartzMouseBaseListenerTestCase(TestCase):
         # Now there are three events
         self.assertEqual(len(listener.events), 3)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (4, 4, 589827))
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (1, 274, 0))
-        third_event = next(inputs.iter_unpack(
+        third_event = next(iter_unpack(
             listener.events[2]))
         self.assertEqual(third_event[2:], (20, 2, 1))
 
@@ -382,11 +387,11 @@ class QuartzMouseBaseListenerTestCase(TestCase):
         # Do we have events
         self.assertEqual(len(listener.events), 2)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (2, 6, 2))
 
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (2, 8, 2))
 
@@ -416,11 +421,11 @@ class QuartzMouseBaseListenerTestCase(TestCase):
         # Do we have events
         self.assertEqual(len(listener.events), 2)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (3, 0, 3))
 
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (3, 1, 2))
 
@@ -450,11 +455,11 @@ class QuartzMouseBaseListenerTestCase(TestCase):
         # Do we have events
         self.assertEqual(len(listener.events), 2)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (2, 0, 600))
 
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (2, 1, 400))
 
@@ -486,7 +491,7 @@ class QuartzMouseBaseListenerTestCase(TestCase):
 
         # The sync marker was added
         self.assertEqual(len(listener.events), 1)
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (0, 0, 0))
 
@@ -522,7 +527,7 @@ class QuartzMouseBaseListenerTestCase(TestCase):
 
         # The sync marker was added
         self.assertEqual(len(listener.events), 1)
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (0, 0, 0))
 
@@ -593,10 +598,10 @@ class AppKitMouseBaseListenerTestCase(TestCase):
         event = mock.MagicMock(return_value=1)
         listener.handle_button(event, 25)
         self.assertEqual(len(listener.events), 2)
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (4, 4, 589827))
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (1, 274, 1))
 
@@ -619,11 +624,11 @@ class AppKitMouseBaseListenerTestCase(TestCase):
         # Check that we have events
         self.assertEqual(len(listener.events), 2)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (3, 0, 600))
 
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (3, 1, 400))
 
@@ -648,15 +653,15 @@ class AppKitMouseBaseListenerTestCase(TestCase):
         # Check that we have events
         self.assertEqual(len(listener.events), 3)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (2, 6, 5))
 
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (2, 8, 5))
 
-        third_event = next(inputs.iter_unpack(
+        third_event = next(iter_unpack(
             listener.events[2]))
         self.assertEqual(third_event[2:], (2, 7, 5))
 
@@ -681,15 +686,15 @@ class AppKitMouseBaseListenerTestCase(TestCase):
         # Check that we have events
         self.assertEqual(len(listener.events), 3)
 
-        first_event = next(inputs.iter_unpack(
+        first_event = next(iter_unpack(
             listener.events[0]))
         self.assertEqual(first_event[2:], (2, 0, 5))
 
-        second_event = next(inputs.iter_unpack(
+        second_event = next(iter_unpack(
             listener.events[1]))
         self.assertEqual(second_event[2:], (2, 1, 5))
 
-        third_event = next(inputs.iter_unpack(
+        third_event = next(iter_unpack(
             listener.events[2]))
         self.assertEqual(third_event[2:], (2, 2, 5))
 
@@ -891,7 +896,7 @@ class AppKitKeyboardListenerTestCase(TestCase):
         event = mock.MagicMock()
         listener.handle_input(event)
         self.assertEqual(len(listener.events), 3)
-        event_info = inputs.iter_unpack(listener.events[1])
+        event_info = iter_unpack(listener.events[1])
         self.assertEqual(next(event_info)[2:], (1, 35, 1))
         mock_get_event_key_code.assert_called_once_with(event)
         mock_get_event_type.assert_called_once_with(event)
@@ -923,7 +928,7 @@ class AppKitKeyboardListenerTestCase(TestCase):
         event = mock.MagicMock()
         listener.handle_input(event)
         self.assertEqual(len(listener.events), 3)
-        event_info = inputs.iter_unpack(listener.events[1])
+        event_info = iter_unpack(listener.events[1])
         self.assertEqual(next(event_info)[2:], (1, 0, 1))
         mock_get_event_key_code.assert_called_once_with(event)
         mock_get_event_type.assert_called_once_with(event)
